Move ValueRefBase into dedicated header file
