+++ UI/EncyclopediaDetailPanel.cpp
@@ -9,6 +9,7 @@
 #include "CUILinkTextBlock.h"
 #include "MapWnd.h"
 #include "../universe/ConditionBase.h"
+#include "../universe/ValueRefBase.h"
 #include "../universe/Encyclopedia.h"
 #include "../universe/Universe.h"
 #include "../universe/Tech.h"
@@ -22,7 +23,6 @@
 #include "../universe/Field.h"
 #include "../universe/Effect.h"
 #include "../universe/Predicates.h"
-#include "../universe/ValueRef.h"
 #include "../Empire/Empire.h"
 #include "../Empire/EmpireManager.h"
 #include "../util/EnumText.h"
