+++ UI/TechTreeWnd.cpp
@@ -14,8 +14,8 @@
 #include "../util/Directories.h"
 #include "../universe/Tech.h"
 #include "../universe/Effect.h"
-#include "../universe/ValueRef.h"
 #include "../universe/Enums.h"
+#include "../universe/ValueRefBase.h"
 #include "../Empire/Empire.h"
 #include "TechTreeLayout.h"
 #include "TechTreeArcs.h"
