+++ parse/ConditionParser2.cpp
@@ -1,8 +1,8 @@
 #include "ConditionParser2.h"
 
 #include "../universe/Condition.h"
-#include "../universe/ValueRef.h"
 #include "../universe/Enums.h"
+#include "../universe/ValueRefBase.h"
 
 #include <boost/spirit/include/phoenix.hpp>
