+++ parse/ConditionParser7.cpp
@@ -3,7 +3,7 @@
 #include "ValueRefParser.h"
 
 #include "../universe/Condition.h"
-#include "../universe/ValueRef.h"
+#include "../universe/ValueRefBase.h"
 
 #include <boost/spirit/include/phoenix.hpp>
