+++ parse/EffectParser2.cpp
@@ -1,9 +1,10 @@
 #include "EffectParser2.h"
 
 #include "ValueRefParser.h"
+#include "../universe/ConditionBase.h"
 #include "../universe/Effect.h"
 #include "../universe/Enums.h"
-#include "../universe/ValueRef.h"
+#include "../universe/ValueRefBase.h"
 
 #include <boost/spirit/include/phoenix.hpp>
