+++ parse/ShipHullsParser.cpp
@@ -9,8 +9,9 @@
 #include "ValueRefParser.h"
 #include "CommonParamsParser.h"
 
-#include "../universe/ValueRef.h"
+#include "../universe/ConditionBase.h"
 #include "../universe/ShipDesign.h"
+#include "../universe/ValueRefBase.h"
 
 #include <boost/spirit/include/phoenix.hpp>
 //TODO: replace with std::make_unique when transitioning to C++14
