+++ parse/ShipPartsParser.cpp
@@ -10,8 +10,9 @@
 #include "ConditionParserImpl.h"
 #include "CommonParamsParser.h"
 
+#include "../universe/ConditionBase.h"
 #include "../universe/ShipDesign.h"
-#include "../universe/ValueRef.h"
+#include "../universe/ValueRefBase.h"
 
 #include <boost/spirit/include/phoenix.hpp>
 //TODO: replace with std::make_unique when transitioning to C++14
