+++ parse/TechsParser.cpp
@@ -6,10 +6,10 @@
 #include "ValueRefParser.h"
 #include "MovableEnvelope.h"
 
-#include "../universe/ValueRef.h"
 #include "../universe/Effect.h"
 #include "../universe/Species.h"
 #include "../universe/Tech.h"
+#include "../universe/ValueRefBase.h"
 #include "../util/Directories.h"
 
 #include <boost/spirit/include/phoenix.hpp>
