+++ server/ServerApp.cpp
@@ -7,6 +7,7 @@
 #include "../combat/CombatLogManager.h"
 #include "../parse/Parse.h"
 #include "../universe/Building.h"
+#include "../universe/ConditionBase.h"
 #include "../universe/Effect.h"
 #include "../universe/Fleet.h"
 #include "../universe/Ship.h"
@@ -18,8 +19,8 @@
 #include "../universe/Species.h"
 #include "../universe/Tech.h"
 #include "../universe/UniverseGenerator.h"
-#include "../universe/ValueRef.h"
 #include "../universe/Enums.h"
+#include "../universe/ValueRefBase.h"
 #include "../Empire/Empire.h"
 #include "../util/Directories.h"
 #include "../util/i18n.h"
@@ -30,6 +31,7 @@
 #include "../util/Order.h"
 #include "../util/OrderSet.h"
 #include "../util/Pending.h"
+#include "../util/Random.h"
 #include "../util/SaveGamePreviewUtils.h"
 #include "../util/SitRepEntry.h"
 #include "../util/ScopedTimer.h"
