+++ universe/Planet.cpp
@@ -1,20 +1,22 @@
 #include "Planet.h"
 
 #include "Building.h"
+#include "ConditionBase.h"
 #include "Fleet.h"
 #include "Ship.h"
 #include "System.h"
 #include "Predicates.h"
 #include "Species.h"
 #include "Universe.h"
-#include "ValueRef.h"
 #include "Enums.h"
+#include "ValueRefBase.h"
 #include "../util/Logger.h"
 #include "../util/GameRules.h"
 #include "../util/OptionsDB.h"
 #include "../util/Random.h"
 #include "../util/Directories.h"
 #include "../util/SitRepEntry.h"
+#include "../util/i18n.h"
 #include "../Empire/Empire.h"
 #include "../Empire/EmpireManager.h"
