+++ universe/Special.cpp
@@ -3,12 +3,13 @@
 #include "ConditionBase.h"
 #include "Effect.h"
 #include "UniverseObject.h"
-#include "ValueRef.h"
+#include "ValueRefBase.h"
 #include "../util/OptionsDB.h"
 #include "../util/Logger.h"
 #include "../util/AppInterface.h"
 #include "../util/CheckSums.h"
 #include "../util/ScopedTimer.h"
+#include "../util/i18n.h"
 
 #include <boost/filesystem/fstream.hpp>
