+++ universe/Universe.cpp
@@ -25,7 +25,7 @@
 #include "Species.h"
 #include "Tech.h"
 #include "Condition.h"
-#include "ValueRef.h"
+#include "ValueRefBase.h"
 #include "Enums.h"
 #include "Pathfinder.h"
 #include "Encyclopedia.h"
